remove a bunch of uses of long (mostly replaced by int32_t)
git-svn-id: svn://localhost/ardour2/branches/3.0@7472 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/route.h
@@ -91,8 +91,8 @@ class Route : public SessionObject, public Automatable, public RouteGroupMember,
 
 	bool set_name (const std::string& str);
 
-	long order_key (std::string const &) const;
-	void set_order_key (std::string const &, long);
+	int32_t order_key (std::string const &) const;
+	void set_order_key (std::string const &, int32_t);
 
 	bool is_hidden() const { return _flags & Hidden; }
 	bool is_master() const { return _flags & MasterOut; }
